temp: simplify Westminster source responsibilities
